requirements.txt
anyio==4.4.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-request-logging==0.7.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
packaging==24.0
pillow==10.2.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
platformdirs==4.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.3
